scripts: Increase USB port numbers
Qemu only has 4 ports that support high speed.
So, when a fifth device is added, it creates an
emulated hub to attach the device. The emulated
hub supports only USB 1.1 and so we get 'speed
mismatch' error whenever a high speed device like
a camera is connected as fifth device. So, change
the default configuration to support more USB2.0
(high speed) and USB 3.0 (superspeed) ports.

Tracked-On: OAM-92919
Signed-off-by: saranya <saranya.gopal@intel.com>