[Revert-me]: Workaround for BT issue in 5.4.121 kernel
With the 5.4.121 kernel BT is not working in host when
android guest is lauched with BT passsthrough and terminated.
In the new firmware download flow, even when firmware is in
operational mode, the firmware version and revision are
checked to see if the firmware in the device is latest.
Because the firmware returns wrong device information, driver
tries to load incorrect firmware file, resulting in failure
in the BT initialisation

Fix: Soft links are created so that ibt-19-16-0* pointing to
ibt-19-0-4*

Patch will be reverted once the firmware changes are available

Tracked-On: OAM-97192
Signed-off-by: Aiswarya Cyriac <aiswarya.cyriac@intel.com>